refactor(admin): extract the subpage and album drawers from PageBuilder (#533) (#547)
Step 3 of #533. Steps 1 and 2 landed in `a3c6611` (#536); step 4
(splitting `admin.css`) is still open, and this PR deliberately does not
touch it — nor the `Listbox` adopt-or-delete decision or a `Field`
wrapper, both of which the issue's comments record as deferred on
purpose.

## What moved

New directory `app/admin/components/page-builder/`:

| File | Lines | What |
| --- | --- | --- |
| `SubpageDrawer.tsx` | 576 | backdrop, container, header/body/footer |
| `AlbumDrawer.tsx` | 250 | the album overlay, fed
`{...getEditingAlbumInfo()}` |
| `SubpagePreview.tsx` | 77 | the drawer's read-only Live Preview tab |
| `AlbumCard.tsx` | 138 | `AlbumCard` + `SortableAlbumCard`, now shared
|
| `sortOptions.tsx` | 43 | `SORT_LABELS` / `SORT_ICONS` / `SORT_OPTIONS`
|
| `types.ts` | 106 | the shared entry/subpage types and the two grid
normalisers |

**No state moved.** `gallery`, `dirty`, `expandedSubpage`, `drawerMode`,
every picker, `markDirty`, `handleSave`, `getEditingAlbumInfo`,
`useScrollLock`, `useUnsavedGuard` and the Escape / ⌘S effects all stay
in `PageBuilder`. The drawers receive callbacks and call them with the
same arguments.

## The number this step was for

| | lines | max indent | lines ≥ 24 |
| --- | --- | --- | --- |
| `PageBuilder.tsx` before | 2,190 | 46 | 628 (29%) |
| `PageBuilder.tsx` after | 1,153 | 22 | 0 |

## Behaviour and appearance

Nothing changes on either. Since nothing renders the panel in CI, the
same two inventory diffs as the previous step were run, and
independently re-run before this PR was opened:

- **CSS class tokens** — old `PageBuilder.tsx` vs. new `PageBuilder.tsx`
plus all extracted components: 119 vs 119, empty diff both ways. No
styling hook lost or invented, so `admin.css` keeps working untouched.
- **User-visible strings** — no label, placeholder, hint, title or
option text differs. The entire string diff is 11 import paths and two
fragments of the new props signature.

Beyond that, the drawer bodies were re-extracted from `HEAD~1`, the
mechanical renames applied (`setExpandedSubpage(null)` → `onClose()`,
`setPickerTarget` → `onPickAlbum`, and so on), whitespace stripped and
compared. `AlbumCard` is byte-identical. The drawers differ in exactly
two places, both Prettier reformatting at the shallower indent: a
redundant paren pair around `{sec.description && …}`, and a `{' '}`
folded into a literal space after `<IconImage size={12} />`.

## Verification

- `npx tsc --noEmit` — 0 errors
- `npm run test:unit` — 820 passed
- `npm run build` — passes
- `npm run lint` — byte-identical to the baseline on `dev`

## What this cannot show

Nothing opened the panel. It is client-side and needs a reachable Immich
instance, and `e2e/` covers the public site only — there is no admin
spec at all. Unchecked by any automated means: drag-and-drop inside the
subpage drawer, the Escape layering, the Edit / Live Preview switch, and
the hero and order pickers opened from inside the album drawer. Two
spots deserve a human's eye specifically: the spacing between icon and
label on the "Pick Hero Image" button (that `{' '}` fold), and the Live
Preview tab, now its own component.

That gap is worth its own issue, as noted on #533 — this step is exactly
the kind of change an admin smoke test would cover.
